类目

+++ src/main/java/com/kunlun/api/service/CategoryServiceImpl.java
@@ -1,12 +1,19 @@
 package com.kunlun.api.service;
 
 import com.alibaba.druid.util.StringUtils;
+import com.github.pagehelper.Page;
+import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.util.StringUtil;
 import com.kunlun.api.mapper.CategoryMapper;
 import com.kunlun.entity.Category;
+import com.kunlun.enums.CommonEnum;
 import com.kunlun.result.DataRet;
 import com.kunlun.result.PageResult;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * @author by hmy
@@ -50,7 +57,7 @@ public class CategoryServiceImpl implements CategoryService {
         if (goodId == null) {
             return new DataRet<>("ERROR", "参数错误");
         }
-        Integer result = categoryMapper.unbindWithGoodId(goodId);
+        Integer result = categoryMapper.unbindGoodId(goodId);
         if (result > 0) {
             return new DataRet<>("解绑成功");
         }
@@ -99,6 +106,13 @@ public class CategoryServiceImpl implements CategoryService {
         if (category.getParentId().intValue() == category.getId().intValue()) {
             return new DataRet<>("ERROR", "修改失败,父id不能为当前记录id");
         }
+        //删除操作
+        if (CommonEnum.UN_NORMAL.getCode().equals(category.getStatus())) {
+            DataRet x = checkStatus(category);
+            if (x != null) {
+                return x;
+            }
+        }
         Integer result = categoryMapper.modify(category);
         if (result > 0) {
             return new DataRet<>("修改成功");
@@ -136,11 +150,16 @@ public class CategoryServiceImpl implements CategoryService {
         if (id == null) {
             return new DataRet<>("ERROR", "参数错误");
         }
+        Category category = categoryMapper.findById(id);
+        DataRet x = checkStatus(category);
+        if (x != null) {
+            return x;
+        }
         Integer result = categoryMapper.deleteById(id);
         if (result > 0) {
             return new DataRet<>("删除成功");
         }
-        return new DataRet<>("ERROR","删除失败");
+        return new DataRet<>("ERROR", "删除失败");
     }
 
 
@@ -154,11 +173,11 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public DataRet<String> updateStatus(String status, Long id) {
         if (id == null || StringUtils.isEmpty(status)) {
-            return new DataRet<>("ERROR","参数错误");
+            return new DataRet<>("ERROR", "参数错误");
         }
-        Integer result=categoryMapper.updateStatus(status, id);
-        if (result==0){
-            return new DataRet<>("ERROR","修改状态失败");
+        Integer result = categoryMapper.updateStatus(status, id);
+        if (result == 0) {
+            return new DataRet<>("ERROR", "修改状态失败");
         }
         return new DataRet<>("修改状态成功");
     }
@@ -174,6 +193,85 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public PageResult findByCondition(Integer pageNo, Integer pageSize, String type, String searchKey) {
+        PageHelper.startPage(pageNo, pageSize);
+        if (StringUtil.isEmpty(type)) {
+            type = null;
+        }
+        if (StringUtil.isEmpty(searchKey)) {
+            searchKey = null;
+        }
+        if (searchKey != null) {
+            searchKey = "%" + searchKey + "%";
+        }
+        Page<Category> page = categoryMapper.findByCondition(searchKey, type);
+        return new PageResult(page);
+    }
+
+
+    /**
+     * 商品批量绑定类目
+     *
+     * @param categoryId
+     * @param goodIdList
+     * @return
+     */
+    @Override
+    public DataRet<String> bindBatch(Long categoryId, List<Long> goodIdList) {
+        if (categoryId == null || goodIdList.size() == 0 || goodIdList == null) {
+            return new DataRet<>("ERROR", "参数错误");
+        }
+        goodIdList.forEach(goodId -> {
+            categoryMapper.unbindGoodId(goodId);
+            categoryMapper.bindCategoryGood(categoryId, goodId);
+        });
+        return new DataRet<>("绑定成功");
+    }
+
+
+    /**
+     * 商品批量解绑
+     *
+     * @param goodIdList
+     * @return
+     */
+    @Override
+    public DataRet<String> unbindBatch(List<Long> goodIdList) {
+        if (goodIdList == null || goodIdList.size() == 0) {
+            return new DataRet<>("ERROR", "参数错误");
+        }
+        Integer result = categoryMapper.unbindCategoryGood(goodIdList);
+        if (result == 0) {
+            return new DataRet<>("ERROR","解绑失败");
+        }
+        return new DataRet<>("解绑成功");
+    }
+
+
+    /**
+     * 判断当前类目是否再使用中
+     *
+     * @param category
+     * @return
+     */
+    private DataRet<String> checkStatus(Category category) {
+        if (category.getParentId() > 0) {
+            //是子类目
+            List<Long> idList = new ArrayList<>();
+            idList.add(category.getId());
+            Integer bindCount = categoryMapper.findCountByCategoryIdList(idList);
+            if (bindCount > 0) {
+                return new DataRet<>("ERROR", "当前类目正在使用中，请先解绑");
+            }
+        } else {
+            //一级类目
+            List<Category> childList = categoryMapper.findByParentId(category.getId());
+            List<Long> idList = new ArrayList<>();
+            childList.forEach(item -> idList.add(item.getId()));
+            Integer bindCount = categoryMapper.findCountByCategoryIdList(idList);
+            if (bindCount > 0) {
+                return new DataRet<>("ERROR", "当前类目子类目正在使用中，请先解绑");
+            }
+        }
         return null;
     }
 }
